backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
lxml>=5.0.0
python-dotenv>=1.0.0
openai>=1.10.0
weasyprint>=60.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
python-multipart>=0.0.6
